Removed some staking circ deps
