+++ src/components/Modals/Send/SendTypes.tsx
@@ -1,0 +1,7 @@
+export enum SendRoutes {
+  Select = '/send/select',
+  Address = '/send/address',
+  Details = '/send/details',
+  Confirm = '/send/confirm',
+  Scan = '/send/scan'
+}
