+++ src/plugins/cosmos/components/modals/Send/views/Confirm.tsx
@@ -20,7 +20,7 @@ import { useHistory } from 'react-router-dom'
 import { Amount } from 'components/Amount/Amount'
 import { MiddleEllipsis } from 'components/MiddleEllipsis/MiddleEllipsis'
 import { useSendFees } from 'components/Modals/Send/hooks/useSendFees/useSendFees'
-import { SendRoutes } from 'components/Modals/Send/Send'
+import { SendRoutes } from 'components/Modals/Send/SendTypes'
 import { TxFeeRadioGroup } from 'components/Modals/Send/TxFeeRadioGroup'
 import { Row } from 'components/Row/Row'
 import { SlideTransition } from 'components/SlideTransition'
