+++ src/plugins/cosmos/components/modals/Staking/views/StakeConfirmRouter.tsx
@@ -6,6 +6,7 @@ import { RouteSteps } from 'components/RouteSteps/RouteSteps'
 import { SlideTransition } from 'components/SlideTransition'
 import { BigNumber } from 'lib/bignumber/bignumber'
 
+import { StakingPath } from '../StakingTypes'
 import { StakeBroadcast } from './StakeBroadcast'
 import { StakeConfirm } from './StakeConfirm'
 
@@ -15,11 +16,6 @@ export type StakingConfirmProps = {
   fiatRate: BigNumber
   apr: string
   onCancel: () => void
-}
-
-export enum StakingPath {
-  Confirm = '/staking/confirm',
-  Broadcast = '/staking/broadcast'
 }
 
 export const stakeConfirmRoutes = [
